Implement a custom TwoPane layout for the LibraryScreen.
When the user clicks on an entry in the library a detail screen is shown as a separate pane. In this screen the user can toggle which of his friends owns the game (this information is necessary for the GameFinder function).
The user can also add new friends using the ActionButton in the NavigationRail.