Implement a basic query-based GameFinder function.
With this function, the user first selects the friends with whom they would like to play together. It is also possible to exclude certain tags in advance. Based on these parameters, the games in question are selected from the database using a query and displayed with a corresponding cover image.