Implement a custom game status.
The goal of this status is to represent the lifecycle of a game in the user's library. The user has also the possibility to add games to a shortlist. The idea behind this list is, that it contains only games that the user would like to play in the short term.